+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Medium performance improvement during startup when unused-public-var linter is disabled.
   - Small performance improvement during startup on unused-public-var calculation parallelizing calculations.
   - Small performance improvement on code actions calcullation.
+  - Add `:use-source-paths-from-classpath` setting defaulting to true, which makes clojure-lsp do not manually discovery source-paths but get from classpath excluding jar files and paths outside project-root. #752 #551
   
 - Editor
   - Fix exception during code actions calculation when in a invalid code of a map with not even key-pairs.